Split expression control flow checking
